pyproject.toml
[tool.poetry]
name = "tg-bot2"
version = "0.1.0"
description = ""
authors = ["0xshodan <hngdean@gmail.com>"]
readme = "README.md"
packages = [{include = "tg_bot2"}]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^2.25.1"
tortoise-orm = "^0.19.3"
environs = "^9.5.0"
loguru = "^0.7.0"
redis = {extras = ["hiredis"], version = "^4.5.5"}
fastapi-admin = {git = "https://github.com/0xshodan/fastapi-admin.git"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
